LIB: fix compilation of sm_round_positive() on non-x86 targets

+++ lib/smmath.hh
@@ -401,7 +401,7 @@ sm_round_positive (float f)
 }
 #else
 inline int
-sm_round_positive (float d)
+sm_round_positive (double d)
 {
   return int (d + 0.5);
 }
